Refactor Playlists VC
